+++ lib/lita/handlers/zendesk.rb
@@ -1,70 +1,48 @@
-require 'base64'
-require 'faraday'
-require 'faraday_middleware'
-
-require 'pp'
+require 'zendesk_api'
 
 module Lita
   module Handlers
     class Zendesk < Handler
 
-      VERSION_URL = 'api/v2'
-      QUERY_SEARCH_PREFIX = 'search.json?query='
-      QUERY_TICKETS_SEARCH = 'search.json?query=type:ticket '
-      QUERY_TICKETS_ALL = 'tickets'
-      QUERY_TICKETS_ESCALATED = 'search.json?query=tags:escalated+status:open+status:pending+type:ticket'
-      QUERY_TICKETS_HOLD = 'search.json?query=status:hold+type:ticket'
-      QUERY_TICKETS_OPEN = 'search.json?query=status:open+type:ticket'
-      QUERY_TICKETS_NEW = 'search.json?query=status:new+type:ticket'
-      QUERY_TICKETS_PENDING = 'search.json?query=status:pending+type:ticket'
-      QUERY_TICKETS_UNSOLVED = 'search.json?query=status<solved+type:ticket'
+      API_VERSION_URL_PATH = 'api/v2'
+      WEB_TICKETS_URL_PATH = 'tickets'
+      QUERY_TICKETS_ALL = 'type:ticket'
+      QUERY_TICKETS_ESCALATED = 'type:ticket tags:escalated status:open status:pending'
+      QUERY_TICKETS_HOLD = 'type:ticket status:hold'
+      QUERY_TICKETS_OPEN = 'type:ticket status:open'
+      QUERY_TICKETS_NEW = 'type:ticket status:new'
+      QUERY_TICKETS_PENDING = 'type:ticket status:pending'
+      QUERY_TICKETS_UNSOLVED = 'type:ticket status<solved'
       QUERY_USERS = 'users'
 
-      config :use_command, types: [TrueClass, FalseClass], default: true
-
       config :subdomain, type: String, required: true
-      config :auth_type, type: String, default: 'password' # or token
-      config :user, type: String, required: true
+      config :username, type: String, required: true
       config :token, type: String, default: ''
       config :password, type: String, default: ''
 
-      def check_client(reload = false)
-        return if @conn && !reload
-        @base_url = base_url
-        @version_url = "#{@base_url}/#{VERSION_URL}"
-        @tickets_url = "#{@base_url}/tickets"
+      config :max_results, type: Integer, default: 10
 
-        if config.auth_type == 'password'
-          @conn = Faraday.new(url: @version_url) do |faraday|
-            faraday.headers['Authorization'] = "Basic #{basic_credentials}"
-            faraday.response :json                    # JSON response
-            faraday.response :logger                  # log requests to STDOUT
-            faraday.adapter  Faraday.default_adapter  # make requests with Net::HTTP
-          end
-        else
-          @conn = Faraday.new(url: @version_url) do |faraday|
-            faraday.response :json                    # JSON response
-            faraday.response :logger                  # log requests to STDOUT
-            faraday.adapter  Faraday.default_adapter  # make requests with Net::HTTP
-          end
-          @conn.basic_auth("#{config.user}/token", config.token) 
+      def client
+        return @_client if @_client
+        Lita.logger.info "#{logger_prefix}Connecting Zendesk Client to #{api_version_url}"
+        @_client = ZendeskAPI::Client.new do |client|
+          client.url = api_version_url.to_s
+          client.username = config.username || config.user
+          client.token = config.token
+          client.password = config.password
         end
       end
 
       def base_url
-        "https://#{config.subdomain.to_s}.zendesk.com"
+        "https://#{config.subdomain}.zendesk.com"
       end
 
-      def basic_credentials
-        Base64.encode64("#{config.user}:#{config.password}").gsub(/\s/,'')
+      def api_version_url
+        uri_join base_url, API_VERSION_URL_PATH
       end
 
-      def zendesk_request(url)
-        check_client
-        if url.index('http') != 0
-          url = "#{@version_url}/#{url}"
-        end
-        @conn.get url
+      def web_tickets_url
+        uri_join base_url, WEB_TICKETS_URL_PATH
       end
 
       # General
@@ -76,7 +54,7 @@ module Lita
 
       route(/^(?:zd|zendesk)\s+search\s+tickets?\s+(\S.*?)\s*$/i, :search_tickets, command: true, help: { 'zd search tickets <QUERY>' => 'returns search results' })
       def search_tickets(response)
-        ticket_search response, QUERY_TICKETS_SEARCH, response.matches[0][0]
+        ticket_search response, response.matches[0][0]
       end
 
       # Ticket Counts
@@ -111,7 +89,7 @@ module Lita
         ticket_count response, QUERY_TICKETS_OPEN, 'open'
       end
 
-      route(/^(?:zd|zendesk)\s+on\s+hold\s+tickets?\s*$/i, :onhold_tickets, command: true, help: { 'zd on hold tickets' => 'returns the count of all on hold tickets' })
+      route(/^(?:zd|zendesk)\s+on\s*hold\s+tickets?\s*$/i, :onhold_tickets, command: true, help: { 'zd on hold tickets' => 'returns the count of all on hold tickets' })
       def onhold_tickets(response)
         ticket_count response, QUERY_TICKETS_HOLD, 'on hold'
       end
@@ -148,71 +126,132 @@ module Lita
         ticket_list response, QUERY_TICKETS_OPEN, 'open'
       end
 
-      route(/^(?:zd|zendesk)\s+list\s+on\s+hold\s+tickets?\s*$/i, :onhold_tickets_list, command: true, help: { 'zd list onhold tickets' => 'returns a list of on hold tickets' })
+      route(/^(?:zd|zendesk)\s+list\s+on\s*hold\s+tickets?\s*$/i, :onhold_tickets_list, command: true, help: { 'zd list on hold tickets' => 'returns a list of on hold tickets' })
       def onhold_tickets_list(response)
         ticket_list response, QUERY_TICKETS_HOLD, 'on hold'
       end
 
       # Ticket Details
 
-      route(/^(?:zd|zendesk)\s+ticket\s+(\d+)\s*$/i, :ticket_details, command: true, help: { 'zd ticket <ID>' => 'returns information about the specified ticket' })
-      def ticket_details(response)
+      route(/^(?:zd|zendesk)\s+ticket\s+(\d+)\s*$/i, :ticket_details_with_comments, command: true, help: { 'zd ticket <ID>' => 'returns information about the specified ticket' })
+      def ticket_details_with_comments(response)
         Lita.logger.info "#{logger_prefix}Processing Zendesk Ticket Details"
-        ticket_id = response.matches[0][0]
-        url = "#{QUERY_TICKETS_ALL}/#{ticket_id}.json"
-        res = zendesk_request url
-        data = res.body
-
-        message = "Ticket #{data['ticket']['id']}: #{@tickets_url}/#{data['ticket']['id']}"
-        message += "\nStatus: #{data['ticket']['status'].upcase}"
-        message += "\nUpdated: " + data['ticket']['updated_at']
-        message += "\nAdded: #{data['ticket']['created_at']}"
-        message += "\nSubject: #{data['ticket']['subject']}"
-        message += "\nDescription:\n-----\n#{data['ticket']['description']}\n-----\n"
-        response.reply message
+        ticket_id = response.matches[0][0].to_i
+        begin
+          ticket = client.ticket.find!(id: ticket_id)
+          response.reply get_text_for_ticket_with_comments(ticket)
+        rescue => e
+          Lita.logger.warn "#{logger_prefix}#{e}"
+          response.reply "Error processing ticket #{ticket_id}"
+        end
       end
 
       private
 
-      def ticket_count(response, url, ticket_type = '')
-        Lita.logger.info "#{logger_prefix}Processing Zendesk Ticket Count"
-        res = zendesk_request url
-        ticket_count = res.body['count']
-        ticket_word  = ticket_count == 1 ? 'ticket' : 'tickets'
-        ticket_desc  = ticket_type == '' ? '' : "#{ticket_type} "
-        response.reply "#{ticket_count} #{ticket_desc}#{ticket_word}."
+      def get_text_for_ticket(ticket, include_description = true)
+        Lita.logger.info "#{logger_prefix}Processing Zendesk ticket details for [#{ticket.id}]"
+        unless ticket.is_a? ZendeskAPI::Ticket
+          raise 'ticket is not a ZendeskAPI::Ticket'
+        end
+        message = "# Ticket #{ticket.id}: #{ticket_url_web(ticket.id)}"
+        message += "\n- Subject: #{ticket.subject}"
+        message += "\n- Status: #{ticket.status}"
+        message += "\n- Updated: #{ticket.updated_at}"
+        message += "\n- Created: #{ticket.created_at}"
+        message += "\n- Requester: #{user_display(ticket.requester)}"  
+        message += "\n- Description: #{ticket.description}" if include_description
+        return message
       end
 
-      def ticket_search(response, url, query)
+      def get_text_for_ticket_comments(ticket)
+        Lita.logger.info "#{logger_prefix}Processing Zendesk ticket comments for [#{ticket.id}]"
+        unless ticket.is_a? ZendeskAPI::Ticket
+          raise 'ticket is not a ZendeskAPI::Ticket'
+        end
+        comments_text = []
+        ticket.audits.each_with_index do |audit,i|
+          if (comment = audit.events.detect {|e| e.type.downcase == 'comment'})
+            author_text = user_display comment.author
+            comment_text = "## Comment: #{author_text}"
+            comment_text += "\n- Created: #{audit.created_at}"
+            comment_text += "\n- Comment: #{comment.body}"
+            comments_text.push comment_text
+          end
+        end
+        return comments_text.reverse.join("\n")
+      end
+
+      def get_text_for_ticket_with_comments(ticket)
+        Lita.logger.info "#{logger_prefix}Processing Zendesk ticket details for [#{ticket.id}] with comments"
+        message = get_text_for_ticket ticket, false
+        comments = get_text_for_ticket_comments ticket
+        if !comments.nil? && comments.length>0
+          message += "\n#{comments}"
+        end
+        return message
+      end
+
+      def ticket_count(response, query, ticket_type = '')
+        Lita.logger.info "#{logger_prefix}Processing Zendesk Ticket Count Query [#{query}]"
+        begin
+          ticket_count = client.search!(query: query).count
+          Lita.logger.info "#{logger_prefix}Ticket count #{ticket_count}"
+          ticket_desc  = ticket_type == '' ? '' : "#{ticket_type} "
+          ticket_word  = ticket_count == 1 ? 'ticket' : 'tickets'
+          response.reply "#{ticket_count} #{ticket_desc}#{ticket_word}."
+        rescue
+          response.reply "A Zendesk error has been encountered."
+        end
+      end
+
+      def ticket_search(response, raw_query)
         Lita.logger.info "#{logger_prefix}Processing Zendesk Ticket Search"
-        url += query
-        res = zendesk_request url
-        tickets = res.body['results']
-        tickets.each do |ticket|
-          response.reply "Ticket #{ticket['id']} is #{ticket['status']}: #{@tickets_url}/#{ticket['id']} - #{ticket['subject']}"
-        end
-        ticket_length = tickets.length
-        ticket_count = res.body['count']
-        ticket_word  = ticket_count == 1 ? 'result' : 'results'
-        response.reply "Listing #{ticket_length} of #{ticket_count} matching #{ticket_word}."
+        tickets = client.search!(query: "#{QUERY_TICKETS_ALL} #{raw_query}")
+        serp_count = reply_tickets response, tickets
+        response.reply "Listing #{serp_count} of #{tickets.count} #{ticket_word(tickets.count)} matching #{raw_query}."
       end
 
-      def ticket_list(response, url, ticket_type = '')
-        Lita.logger.info "#{logger_prefix}Processing Zendesk Ticket List"
-        res = zendesk_request url
-        tickets = res.body['results']
-        tickets.each do |ticket|
-          response.reply "Ticket #{ticket['id']} is #{ticket['status']}: #{@tickets_url}/#{ticket['id']} - #{ticket['subject']}"
-        end
-        ticket_length = tickets.length
-        ticket_count = res.body['count']
-        ticket_word  = ticket_count == 1 ? 'ticket' : 'tickets'
-        ticket_desc  = ticket_type == '' ? '' : "#{ticket_type} "
-        response.reply "Listing #{ticket_length} of #{ticket_count} #{ticket_desc}#{ticket_word}."
+      def ticket_list(response, query, ticket_status = '')
+        Lita.logger.info "#{logger_prefix}Processing Zendesk ticket list query"
+        tickets = client.search!(query: query)
+        serp_count = reply_tickets response, tickets
+        ticket_desc = ticket_status == '' ? '' : "#{ticket_status} "
+        response.reply "Listing #{serp_count} of #{tickets.count} #{ticket_desc}#{ticket_word(tickets.count)}."
       end
 
-      def tickets(count)
+      def reply_tickets(response, tickets)
+        serp_count = 0
+        tickets.each_with_index do |ticket, i|
+          break if i >= config.max_results
+          serp_count += 1
+          response.reply "Ticket #{ticket.id} is #{ticket.status}: #{ticket_url_web(ticket.id)} - #{ticket.subject}"
+        end
+        return serp_count
+      end
+
+      def ticket_url_web(ticket_id)
+        uri_join web_tickets_url.to_s, ticket_id.to_s
+      end
+
+      def ticket_word(count)
         count == 1 ? 'ticket' : 'tickets'
+      end
+
+      def user_display(user)
+        parts = []
+        if user.name.to_s.length > 0
+          parts.push user.name
+        end
+        if user.email.to_s.length > 0
+          if parts.length < 1 || user.email != user.name
+            parts.push "(#{user.email})"
+          end
+        end
+        user_text = parts.join(' ')
+      end
+
+      def uri_join(*args)
+        args.join('/').gsub(/\/\s*\//, '/').gsub(/\/+/, '/').gsub(/^(https?:\/)/i, '\1/')
       end
 
       def logger_prefix